pyproject.toml
[project]
name = "portfolio-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "motor>=3.7.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.42.0",
]
